Remove "Back to top" links
This commit removes them because GitHub has provided the "Table of
contents" on the fixed navigation. When the users scroll through the
README, they will get a fixed navigation on the top left of the
browser's window. So, there is no need for "Back to top" link.

Reference: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-readmes#auto-generated-table-of-contents-for-readme-files